+++ packages/augur-ui/src/assets/styles/_colors_redesign.less
@@ -31,7 +31,8 @@
 
 @color-highlight: #665789;
 @color-bright-highlight: #988FEA;
-@color-highlight-masked: fade(#C4C4C4, 10%);
+@color-highlight-fade: #C4C4C4;
+@color-highlight-masked: fade(@color-highlight-fade, 10%);
 
 @color-positive: #00F1C4;
 @color-neutral: #999999;
