+++ packages/augur-ui/src/modules/common/filter-notice.styles.less
@@ -32,8 +32,8 @@
 
 .red {
   svg {
-    height: 19px;
-    width: 19px;
+    height: @size-19;
+    width: @size-19;
     fill: @color-invalid-primary;
   }
 
@@ -44,9 +44,9 @@
 
 .grey {
   svg {
-    height: 19px;
-    width: 19px;
-    fill: #C4C4C4;
+    height: @size-19;
+    width: @size-19;
+    fill: @color-highlight-fade;
   }
 
   rect, path {
